Added many primitive declarations (over 100).
Replaced the error() function with an lpp_error exception.
The lexer is now responsible for opening/closing the file.
Implemented proper support to handle stdin as one of the input file.
Fixed bug in lexer which would not unget() the operators ending a word.
Transformed the token so the filename and line are saved in each token.
Fixed the detection of floating points, if a period is not followed by a digit, it's a word instead.
Changed the case to lowercase so you need to use "declare" and "to", not "DECLARE" and "TO".
Added a primitive keyword to declare a primitive (opposed to a user procedure.)
Added a program keyword which starts the main() function a la Ada, instead of having instructions floating around all over.
Added a void type so an expression can be set to void instead of some other emptiness (i.e. empty word, empty list).
Changed the use of the lexer and parser to make use of shared pointers.
Got lpp to actually parse logo files. It's not yet outputting anything, but it's ready.
Added a license.cpp so we can print the license on the command line.
Added a few other command line options.
Moved everything within the try/catch.
Prevent the use of "-" more than once.
Fixed the parser so the default expression, procedure body, and program body are now parsed in a second pass.
Added a list in the TO, DECLARE, PRIMITIVE so we can include flags such as INLINE or ARITHMETIC. This will be translated to C++ flags or taken in account to know how to handle a primitive.
Now the parser is capable of handling multiple lexer in a row.
Fixed the TO, DECLARE, etc. end of line detection, we allow WORDs as variable names by only if not at the start of the next line.
Force all the parameters to THING instead of keeping them as WORD/QUOTED when not already a THING.
Added a few missing next_..._token() calls.
Added requirement for an END at the end of all DECLARE, PRIMITIVE, etc. and at the end of a PROGRAM.
Added a special character one can use to define a long and short procedure name without having to define two separate entries.
Added THING and VOID as Unary Expressions.