Fixed the parsing of the MINUS and PLUS operators (i.e. no space behind.)
Fixed the names of the variable functions: [gs]et_vector() was meant to be [gs]et_list().
Implemented the minus(), product() and sum() primitives.
Implemented support for the [:rest] in the parser.cpp and the primitives.
The parser output now takes the function declaration in account which means I can handle the required, optional, and :REST variables. (although I don't yet set these up if missing on the input line.)
Actually added support for the returned value of a function to be the input of another.

Added a new "FUNCTION" and a new "PROCEDURE" flag for declarations.
And I actually save the flags in the declaration (I did not call the set_procedure_flags() yet and it was not even implemented.)
As I was at it, I added the "FUNCTION" and "PROCEDURE" keywords too.
So now you can do: TO ..., FUNCTION ..., or PROCEDURE ... and I check whether you did it right (TO without a flag may be either a function or a procedure, nearly as usual in logo.)
I will use these to know whether a procedure returns a value or not.
If not, it's a new function call.
If it does return a value, it's part of the current argument list.
This is a much better scheme than guessing because of a newline and number of minimum/default/maximum arguments.
Fixed the parse_body(), the call to next_body_token() was in the wrong place.
Fixed the call_function() [not yet taking the FUNCTION/PROCEDURE flags in account, though] so an EOF is properly taken as end of argument list.
Fixed all the expression with a binary operator. I have to skip the operator before reading the rhs.
Fixed the PLUS and MINUS expressions, I also have to skip that operator.
Fixed the parenthesis in expressions, I have to skip the '(' before reading the expression between parenthesis.
Also, the TOK_FUNCTION_CALL I create needs a corresponding declaration and the name has to be in lowercase for now.
Tested with an expression such as -5 + 123 * 2 and got the correct results: 241.