Added "Logo errors".
Started testing for when to return, but this is not yet complete. We can't know whether the child WORD or the top WORD should receive all the arguments... (not without parenthesis) so to replicate Logo we may have a better chance at using the default count after all, at least for intermediates (not the top WORD.)
Renamed the primitives with an introducer and procedures too. That way chances of a clash are much smaller.
Many cleanups.