Added the repeat counter to the context.
Added a primitives.cpp to the compiler. It will implement the "control" for now and "inline" later.
Fixed lexer: the operators (+-*/<>=) characters are taken as part of a word when quoted.
Added support for quoted/function-call for build_list(). These are just like words in this case.
Made sure only primitives can be marked "control".
Made the control primitives inline too.