Create & install the rt library and install the main.cpp too.
Updated the mk to make use of the rt library and installed main.cpp.
Tested the install, it worked.
Added a compare(), logical(), and fixed the function declarations to support std::bind().
Added many more primitives (I even tested a few of them).
Set the :arguments variable in main().
Enhanced the lpp__number class very much.
Changed the test of the "is word an integer or floating point" by using the strtoll() and strtod().
Added the to_integer() and to_float() functions.
Renamed lpp.cpp to context.cpp since it is all about the context. Only the header has everything in it.
Renamed the to_definition() as procedure() since we support many keywords, not just "TO".
Fixed the parsing of the parameters. The min/def/max are all checked, especially for calls within an expression.
Added support for adding the optional parameters when not otherwise defined.
Fixed the optional parameter name in the first argument loop. It's a list, the name (:thing) is the first item.