Added some more primitives.
Implemented the WAIT as an inline primitive.
Added an error code.
Allow for saving a copy of the last thrown error in the context so the ERROR can be implemented.
Implemented the user THROW and CATCH which "talk" to each others as expected, using the standard C++ throw/try/catch mechanism.
Removed the THROW from the list of primitives since it's an inline now.
Fixed the types of variables: we now have a CONTEXT version and a PROCEDURE instead of a LOCAL.
Implemented the TEST and save that result in the context.
Allow first/butfirst/last/butlast of a boolean (as if it were a word).
Fixed the SHOW command so it prints the :REST as expected.
Added an add_prop() to the lpp__value class.
Added the '#' operator which returns the COUNT of the next argument.
Removed all the "inline" attributes except for the WAIT because we started the implementation of inline, but are not going to spend the time on all of those just yet.
Allow words after the END keyword (now that we do not allow lose instructions, it works syntaxically.)
Fixed the procedure declarations (now that they are called "procedure_<name>" and accept a context.)
Allow the get_word() to work against a TOK_LIST if the first element is a word (i.e. declaration) and break up the name.