Implemented the until/while/do.until/do.while control primitives.
Added debug to the test executable so we can more easily find the location of a problem.
Added the TIME (returns current timestamp) and ERRORP (say whether there is an error) primitives.
Prevent the get_primitive_name() from returning a name if the context is not of a primitive.
Added a timestamp and a "caught" flag to the error object.
Fixed the usage of the find_thing(), it may return nullptr so we can't directly access get_value().
Fixed several `if(rest == nullptr)` which should have been `if(rest != nullptr)`.
Fixed the error message, we need to place the e.what() in an std::string or we get a boolean.