Added another 4 primitives.
Fixeed the "backslashed" implementation, it would return an integer instead of a boolean.
Fixed the "beforep" it would try to read thing1/thing2 instead of word1/word2.
Added the "or" to the list of primitives to compile.
Fixed the "member?" definition, it takes 2 parameters.
Fixed the "substring?" definition, it takes 2 parameters.