Added the property list primitives.
Changed the ASCII primitive to UNICODE and the old ASCII became BYTE.
Changed the FIRST/LAST so they return unicode characters when used with words.
The old FIRST/LAST became FIRSTBYTE/LASTBYTE and they return a byte instead of a Unicode character.
Added a TODO.txt with a few ideas.