Implemented the I/O library (open/read/write/close...)
Switch the print/type/show/... to use the I/O library `lpp__write_file()`.
Added debug to the compiler to help in debugging (rare but it happens.)
Added a couple of TODOs about problems that still exist.
Allow for deleting global variables and property lists.
Added/adjusted some errors to use with the I/O library.
Fixed the parser output for boolean, it would write 0 or 1 instead of false or true.
Enhanced some error output in the compiler.
Fixed a bug where a declaration's optional parameters would be compiled twice.