Added the NOTIFY and UNNOTIFY primitives which allows for a callback to receive a call whenever a variable is modified.
Moved the notification from the MAKE primitive to before the set_thing() call otherwise the output can be really confusing!
Cleaned up the procedure() early setup. It should be a lot clearer now, althouh we lost some 'const's.
Fixed the declaration_only in procedure(), it would be set when using the PROCEDURE or FUNCTION keywords.
Removed the trace flag on the procedure. We do not make use of it anymore...
Added info about the fact that a list which includes a keyword breaks (i.e. print [a to b])
Added a script to install dependencies automatically.
Added libboost-dev as a dependency required to compile lpp.
Enhanced a few error messages.